Bugfix Great Britain vaccinated people, vaccinated from int to double as "World" esceeded  2.147.483.647

+++ ISEIRV.cs
@@ -28,7 +28,7 @@ namespace LogicLink.Corona {
         /// <summary>
         /// Number of individuals in the V(accinated) compartment. This value is set manually.
         /// </summary>
-        int Vaccinated { get; set; }
+        double Vaccinated { get; set; }
 
         #endregion
